+++ src/plugins/news/screens/article/selector.js
@@ -1,4 +1,5 @@
 import getBlocksmith from '@triniti/cms/plugins/blocksmith/selectors/getBlocksmith';
+import { getFormValues } from 'redux-form';
 import updateScreenSelector from '@triniti/cms/plugins/ncr/screens/node/selector';
 import schemas from './schemas';
 import { formNames } from '../../constants';
@@ -24,5 +25,6 @@ export default (state, ownProps) => {
   return {
     ...updateScreen,
     blocksmithState,
+    formValues: getFormValues(formNames.ARTICLE)(state),
   };
 };