+++ CHANGELOG.md
@@ -2,8 +2,19 @@
 
 
 ## vN.N.N
+* Implement error logging to logstream for raven and blocksmith errors
+* Set center field on Google Map Block using Geocoding API.
 * facebook embed code considered invalid for facebook post block
-* implement error logging to logstream for raven and blocksmith errors
+* add more detail to social block placeholders.
+* Add `title` prop to image asset picker modal and image asset picker field.
+* Set red error message when article title length is at 75 characters or greater.
+
+
+## v0.12.0
+* Reconnect raven on every `@triniti/raven/PUBLISH_MESSAGE_REQUESTED`.
+* Update raven mqtt to disable reconnect until transformWsUrl supports promises.
+* Use toast alerts for stale data when collaborating
+>>>>>>> release-next
 
 
 ## v0.11.0